bugfix(add-methods): change the measure for executionTime

+++ WebAPI_ASP_Net/Utils/Models/MetricModels/ExecutionTimeMetricModel.cs
@@ -2,6 +2,6 @@
 {
     public class ExecutionTimeMetricModel : IMetricModel
     {
-        public long ExecutionTimeMs { get; set; }
+        public double ExecutionTimeMs { get; set; }
     }
 }
